Moved sink to pkg/
