+++ pkg/sink/stdiosink/colors.go
